pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-commit-gen"  # Nom du package uniformisé
version = "0.1.1"
description = "A tool to generate conventional commit messages using Gemini API"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
dependencies = [
    "google-generativeai>=0.5.0",
    "pygit2>=1.12.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
git-commit-gen = "git_commit_gen.git_commit_gen:main"  # Point d’entrée mis à jour

[tool.hatch.build.targets.wheel]
packages = ["git_commit_gen"]